Add webpack,config files and packages to the setup
