+++ package.json
@@ -1,26 +1,39 @@
 {
-  "name": "awesome-books-es6",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "babel-eslint": "^10.1.0",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-plugin-import": "^2.27.5",
-    "hint": "^7.1.8",
-    "stylelint": "^13.13.1",
-    "stylelint-config-standard": "^21.0.0",
-    "stylelint-csstree-validator": "^1.9.0",
-    "stylelint-scss": "^3.21.0"
-  },
-  "dependencies": {
-    "luxon": "^3.3.0"
-  }
+	"name": "awesome-books-es6",
+	"version": "1.0.0",
+	"description": "",
+	"private": true,
+	"scripts": {
+		"deploy": "gh-pages -d dist",
+		"build": "webpack",
+		"start": "webpack serve --open",
+		"watch": "webpack --watch"
+	},
+	"keywords": [],
+	"homepage": "https://lincoln1883.github.io/Awesome-books-ES6",
+	"author": "",
+	"license": "ISC",
+	"devDependencies": {
+		"@babel/plugin-transform-modules-commonjs": "^7.21.5",
+		"babel-eslint": "^10.1.0",
+		"babel-plugin-syntax-dynamic-import": "^6.18.0",
+		"css-loader": "^6.7.3",
+		"eslint": "^7.32.0",
+		"eslint-config-airbnb-base": "^14.2.1",
+		"eslint-plugin-import": "^2.27.5",
+		"hint": "^7.1.8",
+		"html-webpack-plugin": "^5.5.1",
+		"style-loader": "^3.3.2",
+		"stylelint": "^13.13.1",
+		"stylelint-config-standard": "^21.0.0",
+		"stylelint-csstree-validator": "^1.9.0",
+		"stylelint-scss": "^3.21.0",
+		"webpack": "^5.81.0",
+		"webpack-cli": "^5.0.2",
+		"webpack-dev-server": "^4.13.3"
+	},
+	"dependencies": {
+		"luxon": "^3.3.0",
+		"gh-pages": "^5.0.0"
+	}
 }
