+++ src/index.html
@@ -4,7 +4,6 @@
 		<meta charset="UTF-8" />
 		<meta http-equiv="X-UA-Compatible" content="IE=edge" />
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<link rel="stylesheet" href="index.css" />
 		<title>📖 Awesome Books</title>
 	</head>
 	<body>
@@ -60,6 +59,5 @@
 		<footer>
 			<p>&copy; Copyright Awesome books 2023</p>
 		</footer>
-		<script type="module" src="index.js"></script>
 	</body>
 </html>
