+++ src/index.js
@@ -1,8 +1,9 @@
+import './style.css';
+import { DateTime } from 'luxon';
 import { handleFormSubmit, addBookToList } from './modules/addBook.js';
 import { handleRemoveButtonClick } from './modules/removeBook.js';
 import { handleNavigation, toggleMenu } from './modules/navigation.js';
 import Book from './modules/book.js';
-import showTime from './modules/luxonTime.js';
 
 const formButton = document.querySelector('#submit-button');
 const bookList = document.querySelector('.book-display');
@@ -25,5 +26,20 @@ const main = document.querySelector('main');
 const timeSlot = document.createElement('p');
 const section = document.querySelector('section');
 timeSlot.classList.add('localTime');
+
+// const main = document.querySelector('main');
+// const timeSlot = document.createElement('p');
+// const section = document.querySelector('section');
+// timeSlot.classList.add('localTime');
+
+const showTime = () => {
+	setInterval(() => {
+		const now = DateTime.now();
+		const formattedDateTime = now.toFormat('MMMM dd, yyyy - HH:mm:ss a');
+		timeSlot.textContent = `Local Time: ${formattedDateTime}`;
+	}, 1000);
+
+	main.insertBefore(timeSlot, section);
+};
 
 showTime(main, timeSlot, section);
