+++ src/modules/removeBook.js
