+++ webpack.config.js
@@ -1,0 +1,37 @@
+const path = require('path');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+
+module.exports = {
+	mode: 'development',
+	entry: {
+		index: './src/index.js',
+	},
+	devServer: {
+		static: './dist',
+	},
+	plugins: [
+		new HtmlWebpackPlugin({
+			template: './src/index.html',
+		}),
+	],
+	output: {
+		filename: '[name].js',
+		path: path.resolve(__dirname, 'dist'),
+		clean: true,
+	},
+	optimization: {
+		runtimeChunk: 'single',
+	},
+	module: {
+		rules: [
+			{
+				test: /\.css$/i,
+				use: ['style-loader', 'css-loader'],
+			},
+			{
+				test: /\.(png|svg|jpg|jpeg|gif)$/i,
+				type: 'asset/resource',
+			},
+		],
+	},
+};
